feat: add filtering using nuqs in the Servers page
